Refactor: extract DrawableUtils
